requirements.txt
flask
gunicorn
flask-cors
google-generativeai
load-dotenv